Remove origin from targets of RelatedTo links
RelatedTo links are not directional in the MobSTr traceability
information model, but the Connection class inherently is. Before
this commit, the Connection instance that represented a RelatedTo
link would contain an element both as the origin and as the target.
This led to strange behaviour in the visualisation. Now, the origin
is removed from the list of targets, leading to a clearer
visualisation.