Updated to Eclipse Capra 0.9.0
The MobSTr dataset so far has been based on a special version of
Eclipse Capra 0.8.2. With this commit, the dataset now uses the
current development version of Eclipse Capra and has been updated
to make use of the new interfaces and classes. This enables the
user to use the new features of Eclipse Capra, such as meta data
that can now be stored for trace links, the improved view for
Office files, and support for newer Eclipse versions.